// src/types/database.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
// ...
      files: {
        Row: {
          chapter_id: string
          created_at: string
          extraction_status: string | null
          file_name: string
          file_url: string
          id: string
          text_content: string | null
          text_extracted_at: string | null
        }
        Insert: {
          chapter_id: string
          created_at?: string
          extraction_status?: string | null
          file_name: string
          file_url: string
          id?: string
          text_content?: string | null
          text_extracted_at?: string | null
        }
        Update: {
          chapter_id?: string
          created_at?: string
          extraction_status?: string | null
          file_name?: string
          file_url?: string
          id?: string
          text_content?: string | null
          text_extracted_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "files_chapter_id_fkey"
            columns: ["chapter_id"]
            isOneToOne: false
            referencedRelation: "chapters"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}